Fix test cases
