+++ cmd/nodelabeler/main.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/EdgeNet-project/edgenet/pkg/bootstrap"
 	"github.com/EdgeNet-project/edgenet/pkg/controller/apps/v1/nodelabeler"
-	informers "github.com/EdgeNet-project/edgenet/pkg/generated/informers/externalversions"
 	"github.com/EdgeNet-project/edgenet/pkg/signals"
 	kubeinformers "k8s.io/client-go/informers"
 	"k8s.io/klog"
@@ -40,21 +39,14 @@ func main() {
 		log.Println(err.Error())
 		panic(err.Error())
 	}
-	edgenetclientset, err := bootstrap.CreateEdgeNetClientset("kubeconfig")
-	if err != nil {
-		log.Println(err.Error())
-		panic(err.Error())
-	}
+
 	// Start the controller to provide the functionalities of nodelabeler resource
 	kubeInformerFactory := kubeinformers.NewSharedInformerFactory(kubeclientset, time.Second*30)
-	edgenetInformerFactory := informers.NewSharedInformerFactory(edgenetclientset, 0)
 
 	controller := nodelabeler.NewController(kubeclientset,
-		edgenetclientset,
 		kubeInformerFactory.Core().V1().Nodes())
 
 	kubeInformerFactory.Start(stopCh)
-	edgenetInformerFactory.Start(stopCh)
 
 	if err = controller.Run(2, stopCh); err != nil {
 		klog.Fatalf("Error running controller: %s", err.Error())
