+++ pkg/controller/apps/v1/nodelabeler/controller_test.go
@@ -1,97 +1,171 @@
 package nodelabeler
 
 import (
+	"context"
+	"io/ioutil"
+	"log"
+	"os"
+	"reflect"
 	"testing"
+	"time"
+
+	"github.com/EdgeNet-project/edgenet/pkg/signals"
+	"github.com/EdgeNet-project/edgenet/pkg/util"
+	"github.com/sirupsen/logrus"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	kubeinformers "k8s.io/client-go/informers"
+	"k8s.io/client-go/kubernetes"
+	testclient "k8s.io/client-go/kubernetes/fake"
+	"k8s.io/klog"
 )
 
-func TestStartingController(t *testing.T) {
-
+// The main structure of test group
+type TestGroup struct {
+	nodeObj corev1.Node
 }
 
-// func TestStartingController(t *testing.T) {
-// 	g := testGroup{}
-// 	g.Init()
-// 	// Run the controller in a goroutine
-// 	go Start(g.client)
+var controller *Controller
+var kubeclientset kubernetes.Interface = testclient.NewSimpleClientset()
 
-// 	// Prepare cases
-// 	nodeFR := g.nodeObj
-// 	nodeFR.ObjectMeta = metav1.ObjectMeta{
-// 		Name: "fr.edge-net.io",
-// 		Labels: map[string]string{
-// 			"kubernetes.io/hostname": "fr.edge-net.io",
-// 		},
-// 	}
-// 	nodeFR.Status.Addresses = []corev1.NodeAddress{
-// 		corev1.NodeAddress{
-// 			Type:    "InternalIP",
-// 			Address: "132.227.123.47",
-// 		},
-// 	}
-// 	geolabelsFR := map[string]string{
-// 		"edge-net.io/continent":   "Europe",
-// 		"edge-net.io/state-iso":   "IDF",
-// 		"edge-net.io/country-iso": "FR",
-// 		"edge-net.io/city":        "Paris",
-// 		"edge-net.io/lat":         "n48.860700",
-// 		"edge-net.io/lon":         "e2.328100",
-// 	}
-// 	nodeUS := g.nodeObj
-// 	nodeUS.ObjectMeta = metav1.ObjectMeta{
-// 		Name: "us.edge-net.io",
-// 		Labels: map[string]string{
-// 			"kubernetes.io/hostname": "us.edge-net.io",
-// 		},
-// 	}
-// 	nodeUS.Status.Addresses = []corev1.NodeAddress{
-// 		corev1.NodeAddress{
-// 			Type:    "ExternalIP",
-// 			Address: "206.196.180.220",
-// 		},
-// 	}
-// 	nodeUSToUpdate := nodeUS
-// 	nodeUSToUpdate.Status.Addresses[0].Address = "204.102.228.171"
-// 	geolabelsUSUpdated := map[string]string{
-// 		"edge-net.io/continent":   "North_America",
-// 		"edge-net.io/state-iso":   "CA",
-// 		"edge-net.io/country-iso": "US",
-// 		"edge-net.io/city":        "Seaside",
-// 		"edge-net.io/lat":         "n36.621700",
-// 		"edge-net.io/lon":         "w-121.793500",
-// 	}
+func TestMain(m *testing.M) {
+	klog.SetOutput(ioutil.Discard)
+	log.SetOutput(ioutil.Discard)
+	logrus.SetOutput(ioutil.Discard)
 
-// 	cases := map[string]struct {
-// 		Operation string
-// 		Node      corev1.Node
-// 		Expected  map[string]string
-// 	}{
-// 		"fr":        {"create", nodeFR, geolabelsFR},
-// 		"us update": {"update", nodeUSToUpdate, geolabelsUSUpdated},
-// 	}
+	stopCh := signals.SetupSignalHandler()
 
-// 	for k, tc := range cases {
-// 		t.Run(k, func(t *testing.T) {
-// 			if tc.Operation == "create" {
-// 				g.client.CoreV1().Nodes().Create(context.TODO(), tc.Node.DeepCopy(), metav1.CreateOptions{})
-// 			} else if tc.Operation == "update" {
-// 				g.client.CoreV1().Nodes().Create(context.TODO(), tc.Node.DeepCopy(), metav1.CreateOptions{})
-// 				// Wait for the object to be up to date
-// 				time.Sleep(time.Millisecond * 500)
-// 				updatedNode, _ := g.client.CoreV1().Nodes().Get(context.TODO(), tc.Node.GetName(), metav1.GetOptions{})
-// 				g.client.CoreV1().Nodes().Update(context.TODO(), updatedNode, metav1.UpdateOptions{})
-// 			}
-// 			// Wait for the object to be up to date
-// 			time.Sleep(time.Millisecond * 500)
-// 			node, _ := g.client.CoreV1().Nodes().Get(context.TODO(), tc.Node.GetName(), metav1.GetOptions{})
-// 			if !reflect.DeepEqual(node.Labels, tc.Expected) {
-// 				for actualKey, actualValue := range node.Labels {
-// 					for expectedKey, expectedValue := range tc.Expected {
-// 						if actualKey == expectedKey {
-// 							util.Equals(t, expectedValue, actualValue)
-// 						}
-// 					}
-// 				}
-// 			}
-// 		})
-// 	}
-// }
+	go func() {
+		kubeInformerFactory := kubeinformers.NewSharedInformerFactory(kubeclientset, time.Second*30)
+
+		newController := NewController(kubeclientset,
+			kubeInformerFactory.Core().V1().Nodes())
+
+		kubeInformerFactory.Start(stopCh)
+		controller = newController
+		if err := controller.Run(2, stopCh); err != nil {
+			klog.Fatalf("Error running controller: %s", err.Error())
+		}
+	}()
+
+	os.Exit(m.Run())
+	<-stopCh
+}
+
+func (g *TestGroup) Init() {
+	// Delete the existing Nodes
+	nodeRaw, _ := kubeclientset.CoreV1().Nodes().List(context.TODO(), metav1.ListOptions{})
+	for _, nodeRow := range nodeRaw.Items {
+		kubeclientset.CoreV1().Nodes().Delete(context.TODO(), nodeRow.GetName(), metav1.DeleteOptions{})
+	}
+
+	nodeObj := corev1.Node{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Node",
+			APIVersion: "apps.edgenet.io/v1alpha",
+		},
+		Status: corev1.NodeStatus{
+			Capacity: corev1.ResourceList{
+				corev1.ResourceMemory:           resource.MustParse("3781924"),
+				corev1.ResourceCPU:              resource.MustParse("2"),
+				corev1.ResourceEphemeralStorage: resource.MustParse("51493088"),
+				corev1.ResourcePods:             resource.MustParse("100"),
+			},
+			Allocatable: corev1.ResourceList{
+				corev1.ResourceMemory:           resource.MustParse("3781924"),
+				corev1.ResourceCPU:              resource.MustParse("2"),
+				corev1.ResourceEphemeralStorage: resource.MustParse("51493088"),
+				corev1.ResourcePods:             resource.MustParse("100"),
+			},
+			Conditions: []corev1.NodeCondition{
+				{
+					Type:   "Ready",
+					Status: "True",
+				},
+			},
+		},
+	}
+
+	g.nodeObj = nodeObj
+
+	namespace := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: "default"}}
+	kubeclientset.CoreV1().Namespaces().Create(context.TODO(), namespace, metav1.CreateOptions{})
+}
+
+func TestAssigningGeoLabels(t *testing.T) {
+	g := TestGroup{}
+	g.Init()
+
+	// Create the Paris Node
+	nodeFR := g.nodeObj.DeepCopy()
+	nodeFR.ObjectMeta = metav1.ObjectMeta{
+		Name: "fr.edge-net.io",
+		Labels: map[string]string{
+			"kubernetes.io/hostname": "fr.edge-net.io",
+		},
+	}
+	nodeFR.Status.Addresses = []corev1.NodeAddress{
+		{
+			Type:    "InternalIP",
+			Address: "132.227.123.51",
+		},
+	}
+	geolabelsFR := map[string]string{
+		"edge-net.io/continent":   "Europe",
+		"edge-net.io/state-iso":   "IDF",
+		"edge-net.io/country-iso": "FR",
+		"edge-net.io/city":        "Paris",
+		"edge-net.io/lat":         "n48.860700",
+		"edge-net.io/lon":         "e2.328100",
+	}
+
+	// Create the US Node
+	nodeUS := g.nodeObj.DeepCopy()
+	nodeUS.ObjectMeta = metav1.ObjectMeta{
+		Name: "us.edge-net.io",
+		Labels: map[string]string{
+			"kubernetes.io/hostname": "us.edge-net.io",
+		},
+	}
+	nodeUS.Status.Addresses = []corev1.NodeAddress{
+		{
+			Type:    "ExternalIP",
+			Address: "206.196.180.220",
+		},
+	}
+	geolabelsUS := map[string]string{
+		"edge-net.io/continent":   "North_America",
+		"edge-net.io/state-iso":   "MD",
+		"edge-net.io/country-iso": "US",
+		"edge-net.io/city":        "College_Park",
+		"edge-net.io/lat":         "n38.989600",
+		"edge-net.io/lon":         "w-76.945700",
+	}
+
+	cases := map[string]struct {
+		Node     *corev1.Node
+		Expected map[string]string
+	}{
+		"fr": {nodeFR, geolabelsFR},
+		"us": {nodeUS, geolabelsUS},
+	}
+
+	for k, tc := range cases {
+		t.Run(k, func(t *testing.T) {
+			kubeclientset.CoreV1().Nodes().Create(context.TODO(), tc.Node.DeepCopy(), metav1.CreateOptions{})
+			// g.handler.SetNodeGeolocation(tc.Node.DeepCopy())
+			node, _ := kubeclientset.CoreV1().Nodes().Get(context.TODO(), tc.Node.GetName(), metav1.GetOptions{})
+			if !reflect.DeepEqual(node.Labels, tc.Expected) {
+				for actualKey, actualValue := range node.Labels {
+					for expectedKey, expectedValue := range tc.Expected {
+						if actualKey == expectedKey {
+							util.Equals(t, expectedValue, actualValue)
+						}
+					}
+				}
+			}
+		})
+	}
+}
